# packages/contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "Election": {
        "properties": {
          "contestants": {
            "items": {
              "$ref": "#/components/schemas/Contestant"
            },
            "title": "Contestants",
            "type": "array"
          },
          "date_confidence": {
            "title": "Date Confidence",
            "type": "string"
          },
          "election_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Election Date"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "jurisdiction_id": {
            "title": "Jurisdiction Id",
            "type": "string"
          },
          "last_updated": {
            "format": "date-time",
            "title": "Last Updated",
            "type": "string"
          },
          "majority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Majority"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "seats_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seats Total"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "system": {
            "$ref": "#/components/schemas/ElectionSystem"
          },
          "threshold": {
            "default": 0,
            "exclusiveMaximum": 1.0,
            "minimum": 0.0,
            "title": "Threshold",
            "type": "number"
          }
        },
        "required": [
          "id",
          "jurisdiction_id",
          "name",
          "election_date",
          "date_confidence",
          "system",
          "status",
          "last_updated",
          "contestants",
          "sources"
        ],
        "title": "Election",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Versioned election catalog and probabilistic forecast interface.",
    "title": "Elexion Global API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/elections": {
      "get": {
        "operationId": "elections_v1_elections_get",
        "parameters": [
          {
            "in": "query",
            "name": "jurisdiction",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Jurisdiction"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Election"
                  },
                  "title": "Response Elections V1 Elections Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Elections",
        "tags": [
          "calendar"
        ]
      }
    },
    "/v1/elections/{election_id}": {
      "get": {
        "operationId": "election_detail_v1_elections__election_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "election_id",
            "required": true,
            "schema": {
              "title": "Election Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Election Detail",
        "tags": [
          "forecast"
        ]
      }
    },
    "/v1/elections/{election_id}/backtests": {
      "get": {
        "operationId": "backtests_v1_elections__election_id__backtests_get",
        "parameters": [
          {
            "in": "path",
            "name": "election_id",
            "required": true,
            "schema": {
              "title": "Election Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelComparison"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backtests",
        "tags": [
          "backtests"
        ]
      }
    },
    "/v1/elections/{election_id}/coalitions": {
      "get": {
        "operationId": "coalitions_v1_elections__election_id__coalitions_get",
        "parameters": [
          {
            "in": "path",
            "name": "election_id",
            "required": true,
            "schema": {
              "title": "Election Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoalitionReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Coalitions",
        "tags": [
          "simulations"
        ]
      }
    },
    "/v1/elections/{election_id}/drivers": {
      "get": {
        "operationId": "drivers_v1_elections__election_id__drivers_get",
        "parameters": [
          {
            "in": "path",
            "name": "election_id",
            "required": true,
            "schema": {
              "title": "Election Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Drivers",
        "tags": [
          "drivers"
        ]
      }
    },
    "/v1/elections/{election_id}/forecast": {
      "get": {
        "operationId": "forecast_v1_elections__election_id__forecast_get",
        "parameters": [
          {
            "in": "path",
            "name": "election_id",
            "required": true,
            "schema": {
              "title": "Election Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forecast",
        "tags": [
          "forecast"
        ]
      }
    },
    "/v1/elections/{election_id}/forecast/alternatives/{model_family}": {
      "get": {
        "operationId": "alternative_forecast_v1_elections__election_id__forecast_alternatives__model_family__get",
        "parameters": [
          {
            "in": "path",
            "name": "election_id",
            "required": true,
            "schema": {
              "title": "Election Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "model_family",
            "required": true,
            "schema": {
              "title": "Model Family",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alternative Forecast",
        "tags": [
          "forecast"
        ]
      }
    },
    "/v1/elections/{election_id}/forecasts": {
      "get": {
        "operationId": "forecast_history_v1_elections__election_id__forecasts_get",
        "parameters": [
          {
            "in": "path",
            "name": "election_id",
            "required": true,
            "schema": {
              "title": "Election Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ForecastSnapshot"
                  },
                  "title": "Response Forecast History V1 Elections  Election Id  Forecasts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forecast History",
        "tags": [
          "forecast"
        ]
      }
    },
    "/v1/elections/{election_id}/map-layers": {
      "get": {
        "operationId": "map_layers_v1_elections__election_id__map_layers_get",
        "parameters": [
          {
            "in": "path",
            "name": "election_id",
            "required": true,
            "schema": {
              "title": "Election Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapLayer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Map Layers",
        "tags": [
          "maps"
        ]
      }
    },
    "/v1/elections/{election_id}/mechanics": {
      "get": {
        "operationId": "mechanics_v1_elections__election_id__mechanics_get",
        "parameters": [
          {
            "in": "path",
            "name": "election_id",
            "required": true,
            "schema": {
              "title": "Election Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectionMechanics"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mechanics",
        "tags": [
          "calendar"
        ]
      }
    },
    "/v1/elections/{election_id}/model-comparison": {
      "get": {
        "operationId": "model_comparison_v1_elections__election_id__model_comparison_get",
        "parameters": [
          {
            "in": "path",
            "name": "election_id",
            "required": true,
            "schema": {
              "title": "Election Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelComparison"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Model Comparison",
        "tags": [
          "backtests"
        ]
      }
    },
...
  }
}